Function actions::rcl_action_client_wait_set_get_entities_ready

pub unsafe extern "C" fn rcl_action_client_wait_set_get_entities_ready(
    wait_set: *const rcl_wait_set_t, 
    action_client: *const rcl_action_client_t, 
    is_feedback_ready: *mut bool, 
    is_status_ready: *mut bool, 
    is_goal_response_ready: *mut bool, 
    is_cancel_response_ready: *mut bool, 
    is_result_response_ready: *mut bool
) -> rcl_ret_t

Get the wait set entities that are ready for a rcl_action_client_t. **

The caller can use this function to determine the relevant action client functions

to call: rcl_action_take_feedback(), rcl_action_take_status(),

rcl_action_take_goal_response(), rcl_action_take_cancel_response(), or

rcl_action_take_result_response().

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] wait_set struct where action server services are to be stored

\param[in] action_client an action client to query

\param[out] is_feedback_ready true if there is a feedback message ready to take,

false otherwise

\param[out] is_status_ready true if there is a status message ready to take,

false otherwise

\param[out] is_goal_response_ready true if there is a goal response message ready

to take, false otherwise

\param[out] is_cancel_response_ready true if there is a cancel response message ready

to take, false otherwise

\param[out] is_result_response_ready true if there is a result response message ready

to take, false otherwise

\return RCL_RET_OK if call is successful, or

\return RCL_RET_WAIT_SET_INVALID if the wait set is invalid, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_ACTION_CLIENT_INVALID if the action client is invalid, or

\return RCL_RET_ERROR if an unspecified error occurs. */