Function actions::rcl_action_goal_handle_fini

pub unsafe extern "C" fn rcl_action_goal_handle_fini(
    goal_handle: *mut rcl_action_goal_handle_t
) -> rcl_ret_t

Finalize a rcl_action_goal_handle_t. **

After calling, rcl_action_goal_handle_t will no longer be valid and

rcl_action_server_t will no longer track the goal associated with the goal handle.

After calling, calls to rcl_action_publish_feedback(), rcl_action_publish_status(),

rcl_action_update_goal_state(), rcl_action_goal_handle_get_status(),

rcl_action_goal_handle_is_active(), rcl_action_goal_handle_get_message(), and

rcl_action_goal_handle_get_info() will fail when using this goal handle.

However, the given action server is still valid.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[inout] goal_handle struct to be deinitialized

\return RCL_RET_OK if the goal handle was deinitialized successfully, or

\return RCL_RET_ACTION_GOAL_HANDLE_INVALID if the goal handle is invalid, or */