Function actions::rcl_action_goal_handle_init

pub unsafe extern "C" fn rcl_action_goal_handle_init(
    goal_handle: *mut rcl_action_goal_handle_t, 
    goal_info: *const rcl_action_goal_info_t, 
    allocator: rcl_allocator_t
) -> rcl_ret_t

Initialize a rcl_action_goal_handle_t. **

After calling this function on a rcl_action_goal_handle_t, it can be used to update the

goals state with rcl_action_update_goal_state().

It can also be used to query the state of the goal with

rcl_action_goal_handle_get_message() and rcl_action_goal_handle_is_active().

Goal information can be accessed with rcl_action_goal_handle_get_message() and

rcl_action_goal_handle_get_info().

Goal handles are typically initialized and finalized by action servers.

I.e. The allocator should be provided by the action server.

Goal handles are created with rcl_action_accept_new_goal() and destroyed with

rcl_action_clear_expired_goals() or rcl_action_server_fini().

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[out] goal_handle preallocated, zero-initialized, goal handle structure

to be initialized

\param[in] goal_info information about the goal to be copied to the goal handle

\param[in] allocator a valid allocator used to initialized the goal handle

\return RCL_RET_OK if goal_handle was initialized successfully, or

\return RCL_RET_INVALID_ARGUMENT if the allocator is invalid, or

\return RCL_RET_ACTION_GOAL_HANDLE_INVALID if the goal handle is invalid, or

\return RCL_RET_ALREADY_INIT if the goal handle has already been initialized, or

\return RCL_RET_BAD_ALLOC if allocating memory failed */