Function actions::rcl_action_publish_status

pub unsafe extern "C" fn rcl_action_publish_status(
    action_server: *const rcl_action_server_t, 
    status_message: *const c_void
) -> rcl_ret_t

Publish a status array message for accepted goals associated with an action server. **

This function acts like a ROS publisher and is potentially a blocking call.

\see rcl_publish()

A status array message associated with the action server can be created with

rcl_action_get_goal_status_array().

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] action_server handle to the action server that will publish the status message

\param[in] status_message an action_msgs/StatusArray ROS message to publish

\return RCL_RET_OK if the response was sent successfully, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_ACTION_SERVER_INVALID if the action server is invalid, or

\return RCL_RET_ERROR if an unspecified error occurs. */