Function actions::rcl_action_server_fini

pub unsafe extern "C" fn rcl_action_server_fini(
    action_server: *mut rcl_action_server_t, 
    node: *mut rcl_node_t
) -> rcl_ret_t

Finalize an action server. **

After calling, the node will no longer listen to services and topics for this action server.

(assuming this is the only action server of this type in this node).

After calling, calls to rcl_wait(), rcl_action_take_goal_request(),

rcl_action_take_cancel_request(), rcl_action_publish_feedback(),

rcl_action_publish_status(), rcl_action_send_result(), and

rcl_action_send_cancel_response() will fail when using this action server.

Additionally, rcl_wait() will be interrupted if currently blocking.

However, the given node handle is still valid.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[inout] action_server handle to the action_server to be deinitialized

\param[in] node handle to the node used to create the action server

\return RCL_RET_OK if the action server was deinitialized successfully, or

\return RCL_RET_ACTION_SERVER_INVALID if the action server pointer is null, or

\return RCL_RET_NODE_INVALID if the node is invalid, or

\return RCL_RET_ERROR if an unspecified error occurs. */