Function actions::rcl_action_server_get_goal_handles

pub unsafe extern "C" fn rcl_action_server_get_goal_handles(
    action_server: *const rcl_action_server_t, 
    goal_handles: *mut *mut *mut rcl_action_goal_handle_t, 
    num_goals: *mut usize
) -> rcl_ret_t

Get the goal handles for all goals an action server is tracking. **

A pointer to the internally held array of pointers to goal handle structs is returned

along with the number of items in the array.

The returned handle is made invalid if the action server is finalized, if

rcl_shutdown() is called, or if rcl_action_expire_goals() is called and one or more

goals are expired.

The returned handle is not guaranteed to be valid for the life time of the

action server as it may be finalized and recreated itself.

Therefore, it is recommended to get the handle from the action server using

this function each time it is needed and avoid use of the handle

concurrently with functions that might change it.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] action_server handle to the action server

\param[out] goal_handles is set to the array of pointers to goal handles if successful.

\param[out] num_goals is set to the number of goals in the returned array if successful,

not set otherwise.

\return RCL_RET_OK if successful, or

\return RCL_RET_ACTION_SERVER_INVALID if the action server is invalid, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_ERROR if an unspecified error occurs. */