Function actions::rcl_action_server_goal_exists

pub unsafe extern "C" fn rcl_action_server_goal_exists(
    action_server: *const rcl_action_server_t, 
    goal_info: *const rcl_action_goal_info_t
) -> bool

Check if a goal is already being tracked by an action server. **

Checks whether or not a goal is being tracked in the internal goal array.

The goal state has no effect on the return value.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] action_server handle to the action server

\param[in] goal_info handle to a struct containing the goal ID to check for

\return true if action_server is currently tracking a goal with the provided goal ID, or

\return false otherwise. */