Function actions::rcl_action_take_goal_response

pub unsafe extern "C" fn rcl_action_take_goal_response(
    action_client: *const rcl_action_client_t, 
    response_header: *mut rmw_request_id_t, 
    ros_goal_response: *mut c_void
) -> rcl_ret_t

Take a response for a goal request from an action server using a rcl_action_client_t. **

\todo TODO(jacobperron) blocking of take?

\todo TODO(jacobperron) pre-, during-, and post-conditions for message ownership?

\todo TODO(jacobperron) is this thread-safe?

The caller is responsible for ensuring that the type of ros_goal_response

and the type associate with the client (via the type support) match.

Passing a different type produces undefined behavior and cannot be checked

by this function and therefore no deliberate error will occur.

The caller must provide a pointer to an allocated message for the ros_goal_response.

If the take is successful, this function will populate the fields of ros_goal_response.

The ROS message given by the ros_goal_response void pointer is always

owned by the calling code, but should remain constant during execution of this

function. i.e. Before and after calling rcl_action_send_goal_response() the

ros_goal_response message can change, but it cannot be changed during the call to

rcl_action_send_goal_response().

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] action_client handle to the client that will take the goal response

\param[out] response_header pointer to the goal response header

\param[out] ros_goal_response pointer to the response of a goal request

\return RCL_RET_OK if the response was taken successfully, or

\return RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return RCL_RET_ACTION_CLIENT_INVALID if the action client is invalid, or

\return RCL_RET_ACTION_CLIENT_TAKE_FAILED if take failed but no error occurred

    in the middleware, or

\return RCL_RET_ERROR if an unspecified error occurs. */